Added TypeAwareStreams to properly serializes task generated by REPL

+++ src/test/scala/dev/demo/BaseDemo.scala
@@ -11,19 +11,20 @@ import org.apache.tez.dag.api.TezConfiguration
 
 trait BaseDemo {
 
-  def prepare(jobName:String, inputFiles:Array[String]) {
+  def prepare(jobName: String, inputFiles: Array[String]) {
     val cl = ClassLoader.getSystemClassLoader().asInstanceOf[URLClassLoader]
     val m = classOf[URLClassLoader].getDeclaredMethod("addURL", classOf[URL]);
-	m.setAccessible(true);
-	val file = new File("src/test/resources/mini/").toURI().toURL()
-	m.invoke(cl, file)
-    
-    System.setProperty(TezConstants.GENERATE_JAR, "true");
-    val configuration = new TezConfiguration
-    
-    val fs = FileSystem.get(configuration);
-    fs.delete(new Path(jobName + "_out"))
-    inputFiles.foreach(x => fs.copyFromLocalFile(new Path(x), new Path(x)))
-    
+    m.setAccessible(true);
+    val file = new File("src/test/resources/mini/").toURI().toURL()
+    m.invoke(cl, file)
+
+    if (jobName != null && inputFiles != null) {
+      System.setProperty(TezConstants.GENERATE_JAR, "true");
+      val configuration = new TezConfiguration
+
+      val fs = FileSystem.get(configuration);
+      fs.delete(new Path(jobName + "_out"))
+      inputFiles.foreach(x => fs.copyFromLocalFile(new Path(x), new Path(x)))
+    }
   }
 }
